pyproject.toml
[project]
name = "gmail-fast-mcp"
version = "0.1.0"
description = "Gmail MCP server built with FastMCP"
readme = "README.md"
authors = [
    { name = "Jeff Moore", email = "jmoore@manifold.ai" }
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth>=2.23.0",
]

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "gmail_mcp"
